Renamed and split googleOauth and stripe payment logic
