+++ src/googleOauth.js
@@ -1,7 +1,7 @@
-// TODO: File descriptor
+// Contains logic to login and authenticate users through Google Oauth and Firebase Auth
 import { setSyncStorageValue } from "./chromeStorage.js";
 import { initializeApp } from 'firebase/app';
-import { getFirestore, query, collection, where, getDocs, doc, getDoc, setDoc } from 'firebase/firestore';
+import { getFirestore, doc, getDoc, setDoc } from 'firebase/firestore';
 import { getAuth, onAuthStateChanged, GoogleAuthProvider, signInWithCredential } from "firebase/auth";
 
 const isFirefox = typeof browser !== "undefined";
@@ -17,35 +17,6 @@ const firebaseConfig = {
 
 const app = initializeApp(firebaseConfig);
 const db = getFirestore(app);
-
-// Get products and pricing information from Firestore/Stripe
-async function getProducts() {
-	const q = query(
-		collection(db, 'products'),
-		where('active', '==', true)
-	);
-
-	const querySnapshot = await getDocs(q);
-
-	// for each product, get the product price info
-	const productsPromises = querySnapshot.docs.map(async (productDoc) => {
-		let productInfo = productDoc.data();
-
-		// fetch prices subcollection per product
-		const pricesCollection = collection(productDoc.ref, 'prices');
-		const priceQuerySnapshot = await getDocs(pricesCollection);
-
-		// assume there is only one price per product
-		const priceDoc = priceQuerySnapshot.docs[0];
-		productInfo['priceId'] = priceDoc.id;
-		productInfo['priceInfo'] = priceDoc.data();
-		return productInfo;
-	});
-
-	// 'products' is an array of products (including price info)
-	const products = await Promise.all(productsPromises);
-	return products;
-}
 
 // Get user information from storage or by logging in to Google
 // If localOnly is set to true, the function will not attempt to log in to Google if there is no local information (==in sync storage)
@@ -230,40 +201,4 @@ async function runGoogleOauthAuthentication(action, passedToken, generatedState,
 	}
 
 	return googleOauth;
-}
-
-export async function openStripeCheckout() {
-	const currentUser = await getUser(false);
-	// console.log(currentUser);
-
-	// const products = await getProducts();
-	// console.log(products);
-
-	// // TODO: This must be done without reliance on the name being the same in case it changes
-	// const monthlyPrice = products.find(p => p.name === "One Month of Shuffle+");
-	// const yearlyPrice = products.find(p => p.name === "One Year of Shuffle+");
-
-	// // For testing, assume a monthly subscription
-	// let checkoutSessionData = {
-	//   price: monthlyPrice.priceId,
-	//   // success_url: window.location.origin, // can set this to a custom page
-	//   // cancel_url: window.location.origin   // can set this to a custom page
-	// };
-
-	// const checkoutSessionRef = await addDoc(
-	//   // currentUser is provided by firebase, via getAuth().currentUser
-	//   collection(db, `customers/${currentUser.uid}/checkout_sessions`),
-	//   checkoutSessionData
-	// );
-
-	// // The Stripe extension creates a payment link for us
-	// onSnapshot(checkoutSessionRef, (snap) => {
-	//   const { error, url } = snap.data();
-	//   if (error) {
-	//     // handle error
-	//   }
-	//   if (url) {
-	//     window.location.assign(url);  // redirect to payment link
-	//   }
-	// });
 }
